small bugs for csv

+++ ot2_driver/protopiler/protopiler.py
@@ -197,8 +197,9 @@ class ProtoPiler:
                 ):
                     new_volumes = []
                     for vol in self.resources[resource_key][command.volume]:
-                        new_volumes.append(int(vol))
-
+                        # new_volumes.append(int(vol))
+                        new_volumes.append(float(vol))
+                        
                         command.volume = new_volumes
             if isinstance(command, Mix):
                 if ":[" in command.location:
@@ -369,6 +370,7 @@ class ProtoPiler:
                 self.resources[resource.name] = pd.read_excel(
                     resource.location, header=0
                 )
+        # print(self.resources)
 
     def _reset(
         self,
